[Feature] 全量接入标准动作蒙太奇并完成验收 (Complete Managed Montage Adoption)
## 核心改动
- 迁移剩余 17 个玩法 Ability，完成 20 个原生 Montage 入口的 RateWindow/CancelWindow 托管；保持 GAS 激活、消耗和实际取消权属。
- 保留九参数工厂，以激活前 Setter 配置 Bow/处决混出及 Victim 独立停播权属；修复 Light 换段清理导致的交叉混合丢失。
- 增加仅配置动作与真实动画推进验收，同步迁移测试访问器和实例隔离断言；修正直接测试兼容及敌人冷却截止时间精度。
- 更新稳定架构、C 交接和精简验收归档，关闭 03H9 A/B/C 及父阶段；排入 StanceBreak 红标、Small 击退与按需 Big 倍率配置。
- 增加默认使用 Edit/apply_patch 的可审查编辑规则。

## 验证
- 最终 Development Editor 与非 Editor Win64 Development 构建通过。
- 最终 LightBlend-R1 报告单次 34 项 Automation 全部 Success、0 errors；用户确认最终测试/PIE 通过。
- Unreal MCP 只读完成相关 GA、数据资产、Montage、源动画及效果配置的合并 readback。
- 本次提交前限定路径及暂存区 diff --check 通过，未重复运行未变源码测试。

## 复核
- 独立 Fresh Review 发现的一项 P2 已通过 RED/GREEN 定向修复及 Main 增量复核闭环，无未关闭 P0-P2；未冒称独立 Reviewer 复审修复版本。

## 范围说明
- 仅提交批准的源码、测试及五份文档；排除全部 Content、Config、tmp 和其他用户 WIP，未推送。
- H6/H7 独立债务、Shipping/Test 与打包条件保持原状态；资产 readback 不构成干净资产基线。

+++ Source/PolyQuest/Public/AI/EnemyAIController.h
@@ -294,7 +294,7 @@ private:
 	FGameplayTag StunnedStateTag;
 	FGameplayTag TargetAcquiredEventTag;
 	FGameplayTag TargetLostEventTag;
-	float MeleeAttackCooldownEndTime = 0.0f;
+	double MeleeAttackCooldownEndTime = 0.0;
 	float CachedLeashRadius = 0.0f;
 	bool bHasValidAttackSet = false;
 	bool bHasValidAIProfile = false;
